Problem: Users profile pictures are not used (#315)
Solution: Improve error handling on the edit profile page so that users are aware if something is missing and don't take users back to the previous page if there are errors on the form. Also, don't submit the form when a profile picture is added, submit it only when 'Save changes' button is clicked. Cleanup previous code that tried to solve this issue and comment some parts of the uploader system to make it easier to extend and/or modify. Fix failing tests.